requirements.txt
Flask==3.0.0
pymongo==4.6.0
flask-pymongo==2.3.0
requests==2.31.0
beautifulsoup4==4.12.2
markdownify==0.11.6
python-dotenv==1.0.0
openai==1.3.7
anthropic==0.7.8
feedparser==6.0.10
lxml==4.9.3
python-dateutil==2.8.2
typing-extensions==4.8.0
pydantic==2.5.0
